+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased]
 
+- Update: Change Map selectors `No Route Lines`/`No Tracks` to `Route Lines`/`Tracks`
 - New: Added Track and Attachments columns for the logbook table. Shows if a flight record has a track attached and amount of the attachments. Hidden by default
 
 ## [3.8.0] - 06.09.2025